+++ changelog.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 1.4.1 (TBD)
+## 1.4.1 (2025-06-07)
 
-- Fixed issue where switching to The Judge on a new floor from a weapon with the same base ammo type would break The Judge's ammo UI rendering
 - Added "Preallocate Heap Memory" option
+- Fixed issue where switching to The Judge on a new floor from a weapon with the same base ammo type would break The Judge's ammo UI rendering
 - Fixed Elevator rendering issue caused by "Optimize Visibility Checks" option
 
 ## 1.4.0 (2025-06-06)